backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
yfinance==0.2.37
pandas==2.2.1
scikit-learn==1.4.1.post1
tensorflow==2.18.0
numpy==1.26.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
email-validator
